# share/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: kdb3 API
  version: "1.0"
  description: |
    Document database with SQLite storage, change tracking, and incrementally
    updated materialized views.

    **Trust model:** design documents run SQL against SQLite. Prefer a Bearer
    token and a restricted bind address when not on localhost.
  contact:
    name: kdb3
servers:
  - url: /
    description: This kdb3 instance
...
paths:
  /:
    get:
      tags: [Server]
      summary: Server info
      operationId: getInfo
      responses:
        "200":
          description: Server identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /_cat/dbs:
    get:
      tags: [Server]
      summary: List databases
      operationId: listDatabases
      responses:
        "200":
          description: Database names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
              example: ["blog", "testdb"]
        "401":
          $ref: "#/components/responses/Unauthorized"
  /_uuids:
    get:
      tags: [Server]
      summary: Generate document IDs
      operationId: getUUIDs
      parameters:
        - name: count
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        "200":
          description: Generated IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /_replicate:
    post:
      tags: [Replication]
      summary: Replicate a database (CouchDB-style)
      description: |
        Copies documents from `source` to `target`. Each endpoint is either a
        local database name (`"blog"`) or a remote kdb3 URL
        (`"http://host:8001/blog"`); the object form
        `{"url":"…","token":"…","headers":{…}}` carries auth for a protected
        remote.

        - **One-shot** (default): drains every pending change, then returns a
          summary.
        - **Continuous** (`"continuous": true`): subscribes to the source's
          change feed and keeps the target in sync, returning immediately with a
          `replication_id`.
        - **Cancel**: `{"cancel": true, "replication_id": "…"}` (or the same
          `source`/`target`) stops a continuous replication.

        Revisions are per-database integers, so replication is last-writer-wins
        by document id and preserves document content, not the source `_rev`
        number. Documents whose content already matches are skipped (`no_ops`),
        which also prevents echo loops between bidirectional replications.
      operationId: replicate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplicationRequest"
            examples:
              oneShot:
                summary: One-shot local replication
                value: { source: "blog", target: "blog_backup", create_target: true }
              continuous:
                summary: Continuous replication
                value: { source: "blog", target: "blog_backup", continuous: true }
              remotePull:
                summary: Pull from a remote server
                value: { source: { url: "http://other:8001/blog", token: "secret" }, target: "blog", create_target: true }
              cancel:
                summary: Cancel a continuous replication
                value: { cancel: true, replication_id: "0f1e2d…" }
      responses:
        "200":
          description: Replication summary or continuous handle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplicationResult"
        "400":
          description: Invalid request body
        "404":
          description: Source/target database or replication id not found
        "401":
          $ref: "#/components/responses/Unauthorized"
  /_active_tasks:
    get:
      tags: [Replication]
      summary: List running continuous replications
      operationId: activeTasks
      responses:
        "200":
          description: Active replication tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReplicationTask"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /{db}:
    parameters:
      - $ref: "#/components/parameters/db"
    get:
      tags: [Databases]
      summary: Database stats
      operationId: getDatabase
      responses:
        "200":
          description: Database statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatabaseStat"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [Databases]
      summary: Create database
      operationId: putDatabase
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "412":
          description: Database already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [Documents]
      summary: Create or update a document
      description: |
        Body must be JSON (`Content-Type: application/json`). If `_id` is omitted,
        one is assigned. Updates require the current `_rev`.
      operationId: postDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentWrite"
            example:
              _id: "1"
              title: "hello"
      responses:
        "200":
          description: Saved document meta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentMeta"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "406":
          description: Unsupported Content-Type
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Databases]
      summary: Delete database
      operationId: deleteDatabase
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/Error"
...
  /{db}/_changes:
    get:
      tags: [Changes]
      summary: Change feed
      description: |
        One-shot JSON feed (`feed=normal`, default) or continuous Server-Sent Events
        (`feed=eventsource`). SSE emits one `data:` JSON object per change and
        comment heartbeats about every 25s. Use a client that can send
        `Authorization` (e.g. fetch stream / share/js/follow.js); native EventSource
        cannot set Bearer headers.
      operationId: changes
      parameters:
        - $ref: "#/components/parameters/db"
        - name: since
          in: query
          schema:
            type: integer
            format: int64
            default: 0
          description: Return changes with update_seq greater than this value
        - name: limit
          in: query
          schema:
            type: integer
            default: 1000
          description: Max results; for eventsource, max rows per catch-up batch
        - name: descending
          in: query
          schema:
            type: boolean
            default: false
          description: Newest first when true (ignored for feed=eventsource)
        - name: feed
          in: query
          schema:
            type: string
            enum: [normal, eventsource]
            default: normal
          description: normal = one-shot JSON; eventsource = continuous SSE
      responses:
        "200":
          description: Changes (JSON) or SSE stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChangesResult"
            text/event-stream:
              schema:
                type: string
                description: |
                  SSE frames. Change events are `data: {"update_seq", "id", "rev", "deleted"?}`.
                  Heartbeats are SSE comments (`: `).
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/Error"
...
  /{db}/{docid}:
    parameters:
      - $ref: "#/components/parameters/db"
      - $ref: "#/components/parameters/docid"
    get:
      tags: [Documents]
      summary: Get document
      operationId: getDocument
      parameters:
        - name: rev
          in: query
          schema:
            type: integer
          description: Optional revision to fetch
      responses:
        "200":
          description: Document
          headers:
            E-Tag:
              schema:
                type: string
              description: Document revision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/Error"
    head:
      tags: [Documents]
      summary: Document existence / E-Tag
      operationId: headDocument
      parameters:
        - name: rev
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Exists
          headers:
            E-Tag:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [Documents]
      summary: Create or update document by ID
      operationId: putDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentWrite"
      responses:
        "200":
          description: Saved document meta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentMeta"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/Error"
        "406":
          description: Unsupported Content-Type
        "409":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Documents]
      summary: Delete document
      operationId: deleteDocument
      parameters:
        - name: rev
          in: query
          required: true
          schema:
            type: integer
          description: Current revision (required)
      responses:
        "200":
          description: Tombstone meta
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentMeta"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
  /{db}/_design/{docid}/{view}/_dry_run:
    post:
      tags: [Admin]
      summary: Dry-run draft view SQL against a docs sequence window
      description: |
        Executes unsaved setup/run (and optional select) in an in-memory scratch DB
        attached read-only to the live documents database. Does not write view files.
      operationId: dryRunView
      parameters:
        - $ref: "#/components/parameters/db"
        - $ref: "#/components/parameters/designDocid"
        - $ref: "#/components/parameters/view"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViewAtelierDryRunRequest"
      responses:
        "200":
          description: Dry-run succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewAtelierDryRunResult"
        "400":
          description: Invalid SQL (structured diagnostics)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewAtelierDryRunResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/Error"
        "406":
          description: Unsupported Content-Type
  /{db}/_design/{docid}/{view}/_status:
    get:
      tags: [Admin]
      summary: View catch-up status (lag badge)
      operationId: viewStatus
      parameters:
        - $ref: "#/components/parameters/db"
        - $ref: "#/components/parameters/designDocid"
        - $ref: "#/components/parameters/view"
      responses:
        "200":
          description: View vs database update_seq
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/Error"
  /{db}/_design/{docid}/{view}:
    get:
      tags: [Views]
      summary: Execute view (default select)
      operationId: selectView
      parameters:
        - $ref: "#/components/parameters/db"
        - $ref: "#/components/parameters/designDocid"
        - $ref: "#/components/parameters/view"
        - $ref: "#/components/parameters/stale"
        - $ref: "#/components/parameters/include_docs"
      responses:
        "200":
          description: View select result (shape defined by the view SQL)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Required when the server was started with `-token` or `KDB3_TOKEN`.
        Swagger UI and `/_docs` themselves are public; authorize here before Try it out.
  parameters:
    db:
      name: db
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9_]{1,50}$"
      description: Database name (`[a-z0-9_]`, 1–50 chars, not starting with `_`)
      example: testdb
    docid:
      name: docid
      in: path
      required: true
      schema:
        type: string
        pattern: "^[A-Za-z0-9_]{1,50}$"
      description: Document ID
      example: "1"
    designDocid:
      name: docid
      in: path
      required: true
      schema:
        type: string
      description: Design document name without `_design/` prefix
      example: _views
    view:
      name: view
      in: path
      required: true
      schema:
        type: string
      example: _all_docs
...
    stale:
      name: stale
      in: query
      schema:
        type: boolean
        default: false
      description: If true, read without waiting for the view to catch up
...
  schemas:
    OK:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          example: true
...
    ReplicationRequest:
      type: object
      properties:
        source:
          $ref: "#/components/schemas/ReplicationEndpoint"
        target:
          $ref: "#/components/schemas/ReplicationEndpoint"
        continuous:
          type: boolean
          default: false
          description: Subscribe to the source feed and keep replicating.
        create_target:
          type: boolean
          default: false
          description: Create the target database if it does not exist.
        cancel:
          type: boolean
          default: false
          description: Cancel a running continuous replication.
        replication_id:
          type: string
          description: Identifies the replication to cancel.
        since_seq:
          type: integer
          format: int64
          description: Start from this source update sequence (default 0).
    ReplicationResult:
      type: object
      properties:
        ok:
          type: boolean
        replication_id:
          type: string
        session_id:
          type: string
        continuous:
          type: boolean
        cancelled:
          type: boolean
        source_last_seq:
          type: integer
          format: int64
        docs_read:
          type: integer
          format: int64
        docs_written:
          type: integer
          format: int64
        doc_write_failures:
          type: integer
          format: int64
        no_ops:
          type: integer
          format: int64
          description: Documents skipped because the target already matched.
    ReplicationTask:
      type: object
      properties:
        type:
          type: string
          example: replication
        replication_id:
          type: string
        continuous:
          type: boolean
        source:
          type: string
        target:
          type: string
        docs_read:
          type: integer
          format: int64
        docs_written:
          type: integer
          format: int64
        doc_write_failures:
          type: integer
          format: int64
        checkpointed_source_seq:
          type: integer
          format: int64
        started_on:
          type: integer
          format: int64
          description: Unix timestamp when the replication started.
    ServerInfo:
      type: object
      properties:
        name:
          type: string
          example: kdb3
        version:
          type: object
          properties:
            kdb3:
              type: string
              example: "1.0.0"
            commit:
              type: string
              example: "abc1234"
            sqlite:
              type: string
              example: "3.53.4"
    DatabaseStat:
      type: object
      properties:
        name:
          type: string
        update_seq:
          type: integer
          format: int64
        doc_count:
          type: integer
        deleted_doc_count:
          type: integer
      example:
        name: testdb
        update_seq: 1
        doc_count: 1
        deleted_doc_count: 0
    DocumentMeta:
      type: object
      required: [_id, _rev]
      properties:
        _id:
          type: string
        _rev:
          type: integer
        _deleted:
          type: boolean
    DocumentWrite:
      type: object
      additionalProperties: true
      properties:
        _id:
          type: string
        _rev:
          type: integer
        _deleted:
          type: boolean
    Document:
      type: object
      additionalProperties: true
      required: [_id, _rev]
      properties:
        _id:
          type: string
        _rev:
          type: integer
...
    ViewAtelierDryRunRequest:
      type: object
      properties:
        setup:
          type: array
          items:
            type: string
        run:
          type: array
          items:
            type: string
        select:
          type: object
          additionalProperties:
            type: string
        since:
          type: integer
          format: int64
          default: 0
          description: Sequence window start (exclusive)
        limit:
          type: integer
          default: 300
          description: Max documents in the dry-run window
        select_name:
          type: string
          default: default
        params:
          type: object
          additionalProperties:
            type: string
          description: Values for `${param}` placeholders in the named select
        include_sql:
          type: boolean
          default: false
          description: When true, include generated INSERT changeset SQL
...
    ChangesResult:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              update_seq:
                type: integer
                format: int64
              id:
                type: string
              rev:
                type: integer
              deleted:
                type: boolean
        last_seq:
          type: integer
          format: int64
          description: Highest update_seq in results, or since when empty
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: db_not_found
        reason:
          type: string
          example: database not found
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: unauthorized
            reason: missing or invalid bearer token
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
security:
  - bearerAuth: []
  - {}